==> notebooks/belencsf/00-fetching-social-API.ipynb <==
import requests
import os
from dotenv import load_dotenv
from tqdm import tqdm
import pandas as pd
load_dotenv()

# instances.social API
API_PATH = "https://instances.social/api/1.0/"

def get_instances(next_id=None, **karg):
    params=karg
    if next_id:
        params["min_id"] = next_id
    repsonse = requests.get(
        headers={"Authorization": f'Bearer {os.getenv("AUTH_TOKEN")}'},
        url=API_PATH + "instances/list",
        params=params
        )
    repsonse = repsonse.json()
    return repsonse

def write_df_to_file(df, file_path):
    if not os.path.isfile(file_path):
        df.to_csv(file_path, mode='w', index=False, header=True)
    else:
        df.to_csv(file_path, mode='a', index=False, header=False)

responses = 0 
next_id = None

while tqdm(responses == 0 or next_id):
    file_path = '../../data/instances_eng.csv'
    response = get_instances(next_id=next_id, min_users=10, language="en")
    
    instances_df = pd.DataFrame(response["instances"])
    write_df_to_file(instances_df, file_path)
    responses += instances_df.shape[0]

    next_id = response["pagination"]["next_id"]
